Preliminary work to support MySQL
Currently this does not work, primarily because SQLite is very flexible around the return type of coalesce()
(basically it seems to "just work") whereas MySQL types the entire column based on a common type that everything
can be cast to (in our case, a varchar). That means properties are being returned as strings and, worse, comparisons
within SQL are all nonsense.